scanner/requirements.txt
# Scanner Dependencies

# Core dependencies
python>=3.10,<3.14

# Garak integration (optional but recommended)
garak>=0.9.0

# Model connectors
openai>=1.0.0
anthropic>=0.18.0
huggingface-hub>=0.20.0
transformers>=4.30.0

# HTTP client for API calls
httpx>=0.28.0
aiohttp>=3.9.0

# Configuration management
pyyaml>=6.0.1

# Logging and monitoring
structlog>=24.1.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.10.0

# Optional: Docker support for sandboxing
docker>=7.0.0

